Correct lift calculation
Then total lift is the sum of all elevation changes. If you want to set a dig depth, set the first piece of pipe to 0 length and that negative elevation (depth). Note that this will be applied as a negative to the calculation, which is correct, the total lift should be measured from the water surface.